Fix function parameter value overflow that caused problems when copying data from a GDAL dataset. Problem first appeared in ST_Resize(). Ticket #2188
git-svn-id: http://svn.osgeo.org/postgis/branches/2.0@11058 b70326c6-7e19-0410-871a-916f4a2858ee

+++ raster/rt_core/rt_api.c
@@ -1751,7 +1751,7 @@ int
 rt_band_set_pixel_line(
 	rt_band band,
 	int x, int y,
-	void *vals, uint16_t len
+	void *vals, uint32_t len
 ) {
 	rt_pixtype pixtype = PT_END;
 	int size = 0;
